+++ packages/web-sdk/src/index.ts
@@ -168,3 +168,4 @@ export {
 } from './instrumentations/session';
 
 export { getIgnoreUrls } from './utils/url';
+export { stringifyExternalJson } from './utils/json';