+++ packages/web-sdk/src/instrumentations/console/instrumentation.ts
@@ -1,4 +1,11 @@
-import { allLogLevels, BaseInstrumentation, defaultLogArgsSerializer, LogLevel, VERSION } from '@grafana/faro-core';
+import {
+  allLogLevels,
+  BaseInstrumentation,
+  defaultLogArgsSerializer,
+  LogArgsSerializer,
+  LogLevel,
+  VERSION,
+} from '@grafana/faro-core';
 
 import type { ConsoleInstrumentationOptions } from './types';
 
@@ -7,6 +14,7 @@ export class ConsoleInstrumentation extends BaseInstrumentation {
   readonly version = VERSION;
 
   static defaultDisabledLevels: LogLevel[] = [LogLevel.DEBUG, LogLevel.TRACE, LogLevel.LOG];
+  private errorSerializer: LogArgsSerializer = defaultLogArgsSerializer;
 
   constructor(private options: ConsoleInstrumentationOptions = {}) {
     super();
@@ -15,6 +23,7 @@ export class ConsoleInstrumentation extends BaseInstrumentation {
   initialize() {
     this.logDebug('Initializing\n', this.options);
     this.options = { ...this.options, ...this.config.consoleInstrumentation };
+    this.errorSerializer = this.options.errorSerializer ?? defaultLogArgsSerializer;
 
     allLogLevels
       .filter(
@@ -25,7 +34,7 @@ export class ConsoleInstrumentation extends BaseInstrumentation {
         console[level] = (...args) => {
           try {
             if (level === LogLevel.ERROR && !this.options?.consoleErrorAsLog) {
-              this.api.pushError(new Error('console.error: ' + defaultLogArgsSerializer(args)));
+              this.api.pushError(new Error('console.error: ' + this.errorSerializer(args)));
             } else {
               this.api.pushLog(args, { level });
             }